+++ spec/nanoc/base/identifier_spec.rb
@@ -117,9 +117,30 @@ describe Nanoc::Identifier do
   describe '#=~' do
     let(:identifier) { described_class.new('/foo/bar/') }
 
-    it 'matches on string' do
-      expect(identifier).to be =~ %r{\A/foo/bar}
-      expect(identifier).not_to be =~ %r{\A/qux/monkey}
+    subject { identifier =~ pat }
+
+    context 'given a regex' do
+      context 'matching regex' do
+        let(:pat) { %r{\A/foo/bar} }
+        it { is_expected.to eql(0) }
+      end
+
+      context 'non-matching regex' do
+        let(:pat) { %r{\A/qux/monkey} }
+        it { is_expected.to eql(nil) }
+      end
+    end
+
+    context 'given a string' do
+      context 'matching string' do
+        let(:pat) { '/foo/*/' }
+        it { is_expected.to eql(0) }
+      end
+
+      context 'non-matching string' do
+        let(:pat) { '/qux/*/' }
+        it { is_expected.to eql(nil) }
+      end
     end
   end