+++ NEWS.md
@@ -10,6 +10,7 @@ Enhancements:
 
 * Removed unused options from CLI
 * Added Nanoc::Identifier#without_ext
+* Made Nanoc::Identifier#=~ work with string
 
 Fixes: